requirements.txt
fastapi
uvicorn
tensorflow-cpu
opencv-python
scipy
rasterio
numpy
patchify
requests
earthengine-api
python-multipart
